Rename common_families to common.
