+++ coxeter/families/common.py
